trying to run minecart example

+++ picarx/picarx.py
@@ -39,7 +39,7 @@ class Picarx(object):
         self.camera_servo_pin1 = Servo(PWM(servo_pins[0]))
         self.camera_servo_pin2 = Servo(PWM(servo_pins[1]))   
         self.dir_servo_pin = Servo(PWM(servo_pins[2])) 
-        self.dir_cal_value = int(self.config_flie.get("picarx_dir_servo", default_value=0))
+        self.dir_cal_value = float(self.config_flie.get("picarx_dir_servo", default_value=0))
         self.cam_cal_value_1 = int(self.config_flie.get("picarx_cam_servo1", default_value=0))
         self.cam_cal_value_2 = int(self.config_flie.get("picarx_cam_servo2", default_value=0))
         self.dir_servo_pin.angle(self.dir_cal_value)
@@ -53,7 +53,7 @@ class Picarx(object):
         self.motor_direction_pins = [self.left_rear_dir_pin, self.right_rear_dir_pin]
         self.motor_speed_pins = [self.left_rear_pwm_pin, self.right_rear_pwm_pin]
         self.cali_dir_value = self.config_flie.get("picarx_dir_motor", default_value="[1,1]")
-        self.cali_dir_value = [int(i.strip()) for i in self.cali_dir_value.strip("[]").split(",")]
+        self.cali_dir_value = [float(i.strip()) for i in self.cali_dir_value.strip("[]").split(",")]
         self.cali_speed_value = [0, 0]
         self.dir_current_angle = 0
         for pin in self.motor_speed_pins:
